+++ neoforge/src/main/java/com/sunekaer/sdrp/neoforge/kubejs/SDRPKubeJSIntegration.java
@@ -1,5 +1,4 @@
-//package com.sunekaer.sdrp.integration.kubejs;
-//
+//package com.sunekaer.sdrp.neoforge.kubejs;
 //
 //import dev.architectury.event.EventResult;
 //import dev.architectury.event.events.common.EntityEvent;